+++ htdocs/editor_tab.js
@@ -703,7 +703,7 @@ var editor = function () {
                 var count = 0;
                 var lst = [];
                 function add(items) {
-                    lst.push('&nbsp;');
+                    lst.push(document.createTextNode(String.fromCharCode(160)));
                     lst.push.apply(lst, arguments);
                     ++count;
                 }
@@ -816,12 +816,19 @@ var editor = function () {
         $('span.star', node.element)[0].set_state(i_starred_[node.gistname]);
         $('span.star sub', node.element).text(num_stars_[node.gistname] || 0);
     }
+
+    function make_main_url(notebook, version) {
+        var url = window.location.protocol + '//' + window.location.host + '/main.html?notebook=' + notebook;
+        if(version)
+            url = url + '&version='+version;
+        return url;
+    }
     function tree_click(event) {
         if(event.node.id === 'showmore')
             result.show_history(event.node.parent, false);
         else if(event.node.gistname) {
             if(event.click_event.metaKey || event.click_event.ctrlKey) {
-                var url = window.location.protocol + '//' + window.location.host + '/main.html?notebook=' + event.node.gistname;
+                var url = make_main_url(event.node.gistname, event.node.version);
                 window.open(url, "_blank");
             }
             else {
@@ -932,10 +939,13 @@ var editor = function () {
         save_config: function(k) {
             rcloud.save_user_config(username_, config_, k);
         },
-        load_notebook: function(gistname, version, selroot) {
+        load_notebook: function(gistname, version, selroot, push_history) {
             var that = this;
             selroot = selroot || true;
-            shell.load_notebook(gistname, version, this.load_callback(version, false, selroot));
+            shell.load_notebook(gistname, version,
+                                this.load_callback({version: version,
+                                                    selroot: selroot,
+                                                    push_history: push_history}));
         },
         new_notebook: function() {
             var that = this;
@@ -948,7 +958,7 @@ var editor = function () {
             });
         },
         rename_notebook: function(gistname, newname) {
-            rcloud.rename_notebook(gistname, newname, this.load_callback(null, true, true));
+            rcloud.rename_notebook(gistname, newname, this.load_callback({is_change: true, selroot: true}));
         },
         star_notebook: function(star, opts, k) {
             var that = this;
@@ -978,7 +988,9 @@ var editor = function () {
 
                     if(opts.notebook) {
                         if(opts.make_current)
-                            that.load_callback(opts.version, opts.is_change || false, 'interests') (opts.notebook);
+                            that.load_callback({version: opts.version,
+                                                is_change: opts.is_change || false,
+                                                selroot: 'interests'}) (opts.notebook);
                         else
                             update_notebook_from_gist(opts.notebook, opts.notebook.history, opts.selroot);
                     }
@@ -1028,7 +1040,7 @@ var editor = function () {
         },
         fork_or_revert_notebook: function(is_mine, gistname, version) {
             var that = this;
-            var k = is_mine ? this.load_callback(null, true, true) :
+            var k = is_mine ? this.load_callback({is_change: true, selroot: true}) :
                     function(notebook) {
                         that.star_notebook(true, {notebook: notebook,
                                                   make_current: true,
@@ -1043,8 +1055,13 @@ var editor = function () {
                 $tree_.tree('openNode', node);
             });
         },
-        load_callback: function(version, is_change, selroot, k) {
+        load_callback: function(opts, k) {
             var that = this;
+            var options = $.extend(
+                {version: null,
+                 is_change: false,
+                 selroot: null,
+                 push_history: true}, opts);
             return function(result) {
                 if('error' in result) {
                     k && k();
@@ -1054,13 +1071,24 @@ var editor = function () {
                     throw "Invalid notebook (must have description)";
 
                 config_.currbook = result.id;
-                config_.currversion = version;
+                config_.currversion = options.version;
                 config_.bookuser = result.user.login;
+
+                /*
+                // disabling inter-notebook navigation for now - concurrency issues
+                options.push_history = false;
+                if(options.push_history)
+                    (window.location.search ?
+                     window.history.pushState :
+                     window.history.replaceState)
+                    .bind(window.history)
+                 */
+                window.history.replaceState("rcloud.notebook", null, make_main_url(result.id, options.version));
 
                 var history;
                 // when loading an old version you get truncated history
                 // we don't want that, even if it means an extra fetch
-                if(version)
+                if(options.version)
                     history = null;
                 else
                     history = result.history;
@@ -1068,14 +1096,14 @@ var editor = function () {
                 // get the old history and not the current
                 // this may be the same bug where the latest version doesn't always
                 // show in github
-                if(is_change && shell.is_old_github())
+                if(options.is_change && shell.is_old_github())
                     history.unshift({version:'blah'});
 
                 // check star count if we don't have it in the cache
                 // could always double check here but the thought is that it won't
                 // be terribly out of date
                 var update = function() {
-                    update_notebook_from_gist(result, history, selroot);
+                    update_notebook_from_gist(result, history, options.selroot);
                     that.update_notebook_file_list(result.files);
                 };
                 if(!_.has(num_stars_, result.id))