+++ htdocs/js/bootstrap_utils.js
@@ -3,14 +3,15 @@ var bootstrap_utils = {};
 bootstrap_utils.alert = function(opts)
 {
     opts = _.defaults(opts || {}, {
-        close_button: true
+        close_button: true,
+        on_close: function() {}
     });
     var div = $('<div class="alert"></div>');
     if (opts.html) div.html(opts.html);
     if (opts.text) div.text(opts.text);
     if (opts['class']) div.addClass(opts['class']);
     if (opts.close_button) 
-        div.prepend($('<button type="button" class="close" data-dismiss="alert">&times;</button>'));
+        div.prepend($('<button type="button" class="close" data-dismiss="alert">&times;</button>').click(opts.on_close));
     return div;
 };