+++ htdocs/js/notebook/notebook_controller.js
@@ -29,7 +29,7 @@ Notebook.create_controller = function(model)
                 var filename = file.filename;
                 if(/^part/.test(filename)) {
                     var number = parseInt(filename.slice(4).split('.')[0]);
-                    if(number !== NaN)
+                    if(!isNaN(number))
                         parts[number] = [file.content, file.language, number];
                 }
                 // style..
@@ -206,7 +206,7 @@ Notebook.create_controller = function(model)
                 return {files: _.reduce(changes, xlate_change, {})};
             }
             // not awesome to callback to someone else here
-            k = k || editor.load_callback(null, true, true);
+            k = k || editor.load_callback({is_change: true, selroot: true});
             var k2 = function(notebook) {
                 if('error' in notebook) {
                     k(notebook);