+++ htdocs/js/rcloud_bundle.js
@@ -286,8 +286,30 @@ RCloud.create = function(rcloud_ocaps) {
         };
 
         rcloud.display = {};
+        var cached_device_pixel_ratio;
         rcloud.display.set_device_pixel_ratio = function(k) {
+            cached_device_pixel_ratio = window.devicePixelRatio;
             rcloud_ocaps.set_device_pixel_ratio(window.devicePixelRatio, k || _.identity);
+        };
+        rcloud.display.get_device_pixel_ratio = function() {
+            return cached_device_pixel_ratio;
+        };
+
+        ////////////////////////////////////////////////////////////////////////////////
+        // access the runtime API in javascript as well
+
+        rcloud.api = {};
+        rcloud.api.disable_warnings = function(k) {
+            rcloud_ocaps.api.disable_warnings(k || _.identity);
+        };
+        rcloud.api.enable_warnings = function(k) {
+            rcloud_ocaps.api.enable_warnings(k || _.identity);
+        };
+        rcloud.api.disable_echo = function(k) {
+            rcloud_ocaps.api.disable_echo(k || _.identity);
+        };
+        rcloud.api.enable_echo = function(k) {
+            rcloud_ocaps.api.enable_echo(k || _.identity);
         };
     }
 
@@ -322,6 +344,9 @@ RCloud.create = function(rcloud_ocaps) {
         rcloud.port_notebooks = function(source, notebooks, prefix, k) {
             rcloud_ocaps.port_notebooks(source, notebooks, prefix, k);
         };
+        rcloud.purl_source = function(source, k) {
+            rcloud_ocaps.purl_source(source, k);
+        };
         rcloud.get_completions = function(text, pos, k) {
             return rcloud_ocaps.get_completions(text, pos, function(comps) {
                 if(_.isString(comps))
@@ -346,6 +371,7 @@ RCloud.create = function(rcloud_ocaps) {
         rcloud.session_markdown_eval = function(command, language, silent, k) {
             rcloud_ocaps.session_markdown_eval(command, language, silent, k || _.identity);
         };
+
         rcloud.upload_to_notebook = function(force, on_success, on_failure) {
             on_success = on_success || _.identity;
             on_failure = on_failure || _.identity;
@@ -354,14 +380,21 @@ RCloud.create = function(rcloud_ocaps) {
                 var chunk_size = 1024*1024;
                 var f_size = file.size;
                 var file_to_upload = new Uint8Array(f_size);
+                var bytes_read = 0;
                 var cur_pos = 0;
+                $(".progress").show();
+                $("#progress-bar").css("width", "0%");
+                $("#progress-bar").attr("aria-valuenow", "0");
                 fr.readAsArrayBuffer(file.slice(cur_pos, cur_pos + chunk_size));
                 fr.onload = function(e) {
+                    $("#progress-bar").attr("aria-valuenow", ~~(100 * (bytes_read / f_size)));
+                    $("#progress-bar").css("width", (100 * (bytes_read / f_size)) + "%");
                     if (e.target.result.byteLength > 0) {
                         // still sending data to user agent
                         var bytes = new Uint8Array(e.target.result);
                         file_to_upload.set(bytes, cur_pos);
                         cur_pos += bytes.byteLength;
+                        bytes_read += e.target.result.byteLength;
                         fr.readAsArrayBuffer(file.slice(cur_pos, cur_pos + chunk_size));
                     } else {
                         // done, push to notebook.
@@ -387,6 +420,7 @@ RCloud.create = function(rcloud_ocaps) {
         };
         rcloud.upload_file = function(force, on_success, on_failure) {
             on_success = on_success || _.identity;
+            on_failure = on_failure || _.identity;
             function do_upload(path, file) {
                 var upload_name = path + '/' + file.name;
                 rcloud_ocaps.file_upload.create(upload_name, force, function(result) {
@@ -398,13 +432,20 @@ RCloud.create = function(rcloud_ocaps) {
                     var chunk_size = 1024*1024;
                     var f_size=file.size;
                     var cur_pos=0;
+                    var bytes_read = 0;
+                    $(".progress").show();
+                    $("#progress-bar").css("width", "0%");
+                    $("#progress-bar").attr("aria-valuenow", "0");
                     //initiate the first chunk, and then another, and then another ...
                     // ...while waiting for one to complete before reading another
                     fr.readAsArrayBuffer(file.slice(cur_pos, cur_pos + chunk_size));
                     fr.onload = function(e) {
+                        $("#progress-bar").attr("aria-valuenow", ~~(100 * (bytes_read / f_size)));
+                        $("#progress-bar").css("width", (100 * (bytes_read / f_size)) + "%");
                         if (e.target.result.byteLength > 0) {
                             var bytes = new Uint8Array(e.target.result);
                             rcloud_ocaps.file_upload.write(bytes.buffer, function() {
+                                bytes_read += e.target.result.byteLength;
                                 cur_pos += chunk_size;
                                 fr.readAsArrayBuffer(file.slice(cur_pos, cur_pos + chunk_size));
                             });
@@ -758,14 +799,15 @@ var bootstrap_utils = {};
 bootstrap_utils.alert = function(opts)
 {
     opts = _.defaults(opts || {}, {
-        close_button: true
+        close_button: true,
+        on_close: function() {}
     });
     var div = $('<div class="alert"></div>');
     if (opts.html) div.html(opts.html);
     if (opts.text) div.text(opts.text);
     if (opts['class']) div.addClass(opts['class']);
     if (opts.close_button) 
-        div.prepend($('<button type="button" class="close" data-dismiss="alert">&times;</button>'));
+        div.prepend($('<button type="button" class="close" data-dismiss="alert">&times;</button>').click(opts.on_close));
     return div;
 };
 
@@ -958,9 +1000,15 @@ function create_markdown_cell_html_view(language) { return function(cell_model) 
                 r_result_div.prepend("<pre><code>" + cell_model.content() + "</code></pre>");
             }
 
+            // we use the cached version of DPR instead of getting window.devicePixelRatio
+            // because it might have changed (by moving the user agent window across monitors)
+            // this might cause images that are higher-res than necessary or blurry.
+            // Since using window.devicePixelRatio might cause images
+            // that are too large or too small, the tradeoff is worth it.
+            var dpr = rcloud.display.get_device_pixel_ratio();
             // fix image width so that retina displays are set correctly
             inner_div.find("img")
-                .each(function(i, img) { img.style.width = img.width / window.devicePixelRatio; });
+                .each(function(i, img) { img.style.width = img.width / dpr; });
 
             // capture deferred knitr results
             inner_div.find("pre code")
@@ -1453,7 +1501,7 @@ Notebook.create_controller = function(model)
                 var filename = file.filename;
                 if(/^part/.test(filename)) {
                     var number = parseInt(filename.slice(4).split('.')[0]);
-                    if(number !== NaN)
+                    if(!isNaN(number))
                         parts[number] = [file.content, file.language, number];
                 }
                 // style..
@@ -1630,7 +1678,7 @@ Notebook.create_controller = function(model)
                 return {files: _.reduce(changes, xlate_change, {})};
             }
             // not awesome to callback to someone else here
-            k = k || editor.load_callback(null, true, true);
+            k = k || editor.load_callback({is_change: true, selroot: true});
             var k2 = function(notebook) {
                 if('error' in notebook) {
                     k(notebook);